% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{const}
\alias{const}
\title{Help function}
\usage{
const(gamma)
}
\arguments{
\item{gamma}{A vector.}
}
\description{
This function implements the algorithm to calculate the PAN estimate for a given tuning parameter path
specified by maximum value and step size.
}
