% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PAN.bootstrap}
\alias{PAN.bootstrap}
\title{Bootstrap}
\usage{
PAN.bootstrap(
  X.train,
  lambdaMax = 10,
  lambdaMaxNeg = NULL,
  stepSize = 1,
  stepSizeNeg = NULL,
  bootstrap.samples = 2000
)
}
\arguments{
\item{X.train}{An n x p matrix for the training set. The number of
observations (rows) must be larger than the number of variables (columns).}

\item{lambdaMax}{A scalar. The absolute value of the maximum tuning parameter.}

\item{stepSize}{A scalar. The step size (either positive or negative) for the tuning parameter starting from 0.}

\item{X.test}{An n x p matrix for the test set. The number of
observations (rows) can be one or more.}
}
\description{
This function implements the algorithm to calculate the PAN estimate for a given tuning parameter path
specified by maximum value and step size.
}
\examples{
cat_function()
}
